Add the <Validator> class
Added the class <Validator> located in validator.py, that represents a
statement related to a thesis that we want to prove or disprove. The
statement is represented by a function that returns the epistemological
status of that statement and its relation to the thesis is represented
by a string ("implied", "implying", "equivalent", "opposite",
"contradictory"). The Validator.status method returns the
epistemological status of the thesis given the status of the statement.

For example:
- if the statement is "implied" by the thesis and its status is
    "false", than the status of the thesis is "false" because a "true"
    thesis cannot imply a "false" statement.

- If the statement is "implying" the thesis and its status is "true",
    than the thesis is "true".

- If the status of the statement is "unknown", than the status of the
    thesis is "unknown".

Added a function <validate> that takes an arbitrary number of
<Validator> instances and executes them untill one of them returns
"true" or "false".